% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chat_prompts_js}
\alias{chat_prompts_js}
\title{Generate JavaScript for suggested prompt clicks}
\usage{
chat_prompts_js(container_id)
}
\arguments{
\item{container_id}{The DOM ID of the container element}
}
\value{
A shiny tags$script element
}
\description{
Creates the JS code to handle clicking on prompt chips.
}
\keyword{internal}
